requirements.txt
altair==5.3.0
attrs==23.2.0
blinker==1.7.0
boto3==1.34.84
botocore==1.34.84
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fsspec==2024.3.1
getenv==0.2.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
idna==3.7
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
psycopg==3.1.18
psycopg-binary==3.1.18
pyarrow==15.0.2
PyAthena==3.6.0
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
s3transfer==0.10.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.29
streamlit==1.33.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
